requirements.txt
# Production dependencies
flask>=3.0.0
wtforms>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pysmb>=1.2.9
gunicorn>=21.0.0
